build: add environment files for x86_64 and aarch64
Previously we did not do any microarch optimizations, now we target Nehalem-era
x86_64, tuned for Broadwell or newer microarchitectures.

On the ARM side, we are starting with ARMv8.0-A as baseline, but tuned for
Neoverse N1, which the Ampere Altra core, used by most cloud providers,
is derived from.

In addition, upgrade to -D_FORTIFY_SOURCE=3, which is meaningful in glibc
2.36 and later.