add CRON jobs to periodically check packages for new versions
This PR add's two CRON jobs because depending on which service we use to check for new packages the updater performs differently.

When using GitHub to check for updates the updater uses a graphql API and a single query compared with https://release-monitoring.org/ which needs to call their API for every package we are checking.  Therefore, checking via GitHubs API can be done more frequently.

The CRONS from this PR  will query GitHub every hour and https://release-monitoring.org/ daily.

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>