Add a Redis package.
This applies two patches, one was taken from the Alpine redis distro (redis.confd.patch),
and the other was taken from the Docker image of redis, which is required to disable
protected mode. Protected mode prevents redis from listening on an external port by
default, but we need that capability in containers.

This also adds 3 (overly broad) CVEs to the initial secfixes entry. These aren't applicable.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>