Add the PyYAML package.
This is required now for Ruby 3.2. Prior versions of Ruby bundled this directly,
but now it's an extra dependency.

For some reason I needed to create the empty yaml directory for builds to succeed.
Without it, I ran into issues. I think some of the files in the yaml source tree conflicted.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>