glibc: disable gettext for C.UTF-8 locale and ensure stdio locking is consistent across threads
When C.UTF-8 support was added to glibc, they forgot to disable gettext translations for
the C.UTF-8 locale as they do for the C (ASCII) locale.  This patch should be sent upstream.

When accessing stdio streams, always do locking/unlocking.  This fixes stdio access to be
POSIX-conformant and removes a flawed test.  This patch should also be sent upstream, but
kept if rejected.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>