push packages to wolfi-production-registry-destination directly
the current approach with using a source bucket + reconcilation is not really getting us anything
useful, but it *is* getting us a 2 hour waiting period on builds while the repository reconciles.

we should move to using wolfictl to reconcile the index and upload artifacts directly, but that
requires further discussion.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>