enable so name CI check as a secondary CI check + add hello-wolfi package we can use as a canary package when verifying changes
This is a seconod option (related to #813) to modifying the existing CI check in case we see unexpected behaviour.

The downside is we will perform duplicate builds for a time while we verify further.  So this is a short term option, seen as a step to adding to the main CI build but reducing impact to contributors.

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>