Add systemd package, and add it as a dependency to fluent-bit.
Fluent-bit has a systemd plugin that's required for the default helm chart.
We comfigured it in our build, but didn't actually have systemd-dev so it ignored that part, making
our fluent-bit image crash in a normal helm install.

I've verified that this fixes our image and the helm tests now pass!

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>