Fix versioning scheme for cluster-autoscaler packages.
This is built out of a monorepo that does a bunch of different separate releases.
They release a package for cluster-autoscaler, and then also a chart. And for some
reason they use different versioning schemes, and I picked the wrong ones.

This fixes that, then withdraws the old packages so the correct ones can get installed.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>