Bump the `wolfi-baselayout` epoch
This bumps the `wolfi-baselayout` epoch so that we will start to have a `builddate` in this package that is not the unix epoch.

This should enable us to stop "stamping" our builds with `SOURCE_DATE_EPOCH={now}` and get lower churn images from Wolfi packages.

All new packages already build with the epoch derived from their melange config, so this will trigger the new "build date epoch" calculation in `apko` downstream to start using this timestamp as the new ~floor instead of the unix epoch.