build flags: remove -Wextra and -Wtrampolines
the -Wtrampolines flag is GCC-specific and is intended to issue warnings about the use of a very
rarely used GCC feature that allows dynamically-generated trampolines at runtime to be created.

the only time this happens is when the GCC nested functions extension is used, which is basically
never.  this conflicts with programs built using Clang with -Werror, such as those built by Bazel
and MSBuild.

the -Wextra flag enables a lot of extra warnings that are not enabled by -Wall.  this is useful
for developers, who are reading their build logs, but not useful for packagers who are just
building software, and most likely are only reading build logs on failures.  the -Wextra flag
also causes problems with Bazel and MSBuild for the same reason as -Wtrampolines, the extra
warnings conflict with -Werror.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>