fix: ensure corepack cli available on path
First, the wildcard expansion didn't work on the
original link command, but we actually only need
the corepack shim anyway.

Secondly, we now link directly to corepack.js instead
of the shim, due to how the shim tries to require
(relative path) the dist anyway. We can bypass easy enough