gnutls: fix underlinking
When building with -Wl,-z,now, all symbols must be resolved at link time, as lazy
binding is disabled.  This provides better performance and also provides protections
when combined with -z relro against segments being overwritten at runtime.

However, there is a bug in the GNUTLS build, where the p11-kit symbols are not
linked by the pkg-config file normally, allowing the dependency to be resolved
through lazy binding instead.  Fix this by adding a Requires relationship on the
p11-kit package in the pkg-config file.

Signed-off-by: Ariadne Conill <ariadne@dereferenced.org>