`ruby3.2-validate_url`: unbreak at HEAD.
This fixes an issue with `ruby3.2-validate_url` where the `git clone {url} .` was failing at HEAD because the directory isn't empty (it has `lost+found`).

To address this, I played around with a few options, but ended up adapting a forked version of our current `git-checkout` pipeline.

This also moves 1.0.15 one commit ahead to let us drop the `sed` command, since they updated the upstream repo with a superset of this change.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>