Create more virtual java packages.
Today we have a virtual `default-jvm` package, which installs its JRE, and sets up that JRE as the default.

The above package is of somewhat limited use for us in images (currently) because we track LTS and we don't have a similar virtual that JUST tracks LTS.  So the first piece of this change is introducing a new LTS virtual
`default-jvm-lts` which is provided by the same packages, but only on LTS versions (8, 11, 17).

We are also missing a virtual package that aggregates the JDK into a single logical package.  So the next two pieces of this change are to introduce a similar concept to the above for installing a default JDK (and a similar LTS form)
called `default-jdk` and `default-jdk-lts` respectively.  These are provided by a new empty subpackage, which has a runtime dependency on the appropriately versioned `default-jvm` and `openjdk-X` packages.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>