`gitlab-shell`: new package
This change adds a new package for the `gitlab-shell` project.  The build process is loosely based on the upstream `Makefile` and the downstream CNG `Dockerfile`:
https://gitlab.com/gitlab-org/build/CNG/-/blob/master/gitlab-shell/Dockerfile?ref_type=heads#L27-50

I checked all of the binaries we build in my local APK vs. the binaries in the latest docker image and they are all 5+MB smaller.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>