Remove subpackages that use if
This is only used in a single place to make openssl configurable in a
way that is no longer used, so this just adds complexity. Drop it so we
can drop the logic in melange for handling subpackage ifs.

This also drops the fips build option since it's no longer used and just
hardcodes the no-fips option that is default.

Doing this also removes the only use of top-level build options, which
would mean we could also remove that.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>