Support incremental package uploads
Currently, we build any out of date packages as part of this workflow.
If the build step of the workflow failsf or any package, we don't
proceed to the "upload" job, which means we don't upload any of the
packages that were successfully built.

This change splits the "upload" job into "upload-packages" and
"upload-index".

The "upload-packages" job will always run. It signs the packages and
indexes, but only uploads the packages to GCS; the indexes get uploaded
to GHA Artifacts.

The "upload-index" job only runs if the "build" job succeeded. It
downloads the indexes from GHA Artifacts and uploads them to GCS.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>