Remove wget from default busybox config.
Wget is a http/https client.  We have decided to remove it from
the default busybox build as simply good hygiene.

Images that need a command line http client have the following options:
 * add curl package
 * add wget package
 * use busybox-full package instead of busybox.

Also note that ssl config option is dropped.  This is because
wget is the only thing that was pulling it in.