Add JPEG2000 support and fix gdal version
The cmake build system adds support for plugins when it can detect them in the build system. Adding openjpeg-dev will enable JPEG2000 support.

Switching from git-checkout to fetch resolves a separate issue in which the version (check with $gdal-config --version) returns the version with a "dev" suffix appended,, for instance "3.8.3dev" instead of "3.8.3". The gdal build setup adds "dev" even when the build is marked as a release if it detects a hidden git directory. When the .git directory is not present, we see the correct version string. https://github.com/OSGeo/gdal/blob/master/cmake/helpers/GdalVersion.cmake#L43-L47



Signed-off-by: John Osborne <josborne@chainguard.dev>