vault-1.14: add back the now missing `go mod tidy`
Previously vault-1.14 had a go/bump pipeline, which has a side effect
of calling `go mod tidy`. This step appears to be required, as the
build fails with go toolchain requesting to perform `go mod tidy`
prior to being able to complete the `go generate` steps.

Add back `go mod tidy` as the first pipeline action. It's either
harmless, or unbreaks the build.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@chainguard.dev>