glibc,libxcrypt: move libcrypt1 package from glibc to libxcrypt
This will be required with next glibc release, but was possible to do
for a while now. All other distros have transitioned to using
libxcrypt already. I have double checked the abi compatibility with
abi-compliance-checker. Separately, once this lands, we might want to
bump epoch on all the reverse depends for them to gain stronger crypt
support.

(See
https://apk.dag.dev/https/packages.wolfi.dev/os/aarch64/APKINDEX.tar.gz@etag:2edafbb750de4d6487204c1cd6110c7b/APKINDEX?&depend=so%3Alibcrypt.so.1
for packages to bump).

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@chainguard.dev>