Use wolfictl build for postsubmit builds
This gives us concurrent package builds, which speeds things up a bit.

We can revert this if things don't work as expected, but I've also moved
the current definition to build-old.yaml in case we want to manually
invoke it down the line. We may want to remove that once we've decided
this will stick.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>