Restore python-3.11 provide python-3 and python3 at 3.11.999
This is present here for historic reasons.  Per a slack thread,
older 3.11 versions of python (and 3.8, 3.9, 3.10) provided
  python-3 = 3.X.999
  python3 = 3..999

(Where X = the minor version).

The reason for its presense was that there was no melange substitution
for full-version available and the automatic update pull requests
could not handle changing the provides here.

Note that the change here uses '3.11.999' rather than
'${{vars.pyversion}}.999' because an attempt to do so failed melange
build with:

 failed to read package config python-3.11.yaml:
   failed to apply replacement to provides
     "python3=${{vars.pyversion}}.999":
     variable vars.pyversion not defined

user-defined variables are not available there.