Drop the bash dependency from dependency-track
dependency-track does not need bash. I suspect that the bash
runtime dependency was there to have a shell to run the test in.
Either that, or to run the image entry point.

The image entry point in
images/dependency-track/config/template.apko.yaml
is currently 'sh -c "exec java ${JAVA_OPTIONS}"'. Interestingly,
because bash doesn't actually even provide /bin/sh, it was
busybox which was adding that.