Add test/daemon-check-output pipeline and use it.
I recently had added a test to keycloak and guacamole-server
which would run a the daemon and then look in output for
some given strings.

This adds a reusable piepline to do that which is implemented in shell.
Ultimately it'd be nicer to have this in melange proper without any
shell dependencies.