Attempt to fix gcsfuse read-only issues
Current theory is:

A post-submit partially fails and we upload the packages that were built
but don't update the APKINDEX.

On the next post-submit, we stat those files via gcsfuse, but we get a
file not found, even though those files do exist. For some reason, we
think gcsfuse is too slow or something.

Since they don't appear exist, we attempt to build them, but then get a
read-only filesystem failure because that file does exist and is a
gcsfuse symlink, which means we can't open for write.

This change does two things:

1. Replace the file glob with `find` if there are just too many files.
2. Do a 30 second sleep after we start up gcsfuse.

We're going to try to drop gcsfuse entirely and make wolfictl build do a
HEAD request against the bucket, at which point both these hacks can be
deleted.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>