Be safer with here docs in docker-run and lint-world actions.
This allows the user of docker-run action to not have to worry
that their document will be parsed by 2 shells.

from https://manpages.ubuntu.com/manpages/noble/en/man1/bash.1.html

> If any part of word is quoted, the delimiter  is the  result of quote
> removal on word, and the lines in the here-document are not expanded.
> If word is unquoted, all lines of the here-document are subjected to
> parameter  expansion, command  substitution,  and  arithmetic
> expansion,  the  character sequence \<newline> is ignored, and \ must be
> used to quote the characters \, $, and `.

What that sums up to is that there is less interpretation
on the here document when using `<<"EOF"` than when using `<<EOF`.