llvm-18: fix naming
As per:

https://discourse.llvm.org/t/cross-compile-against-sysroot-of-older-linux-distribution/69927

LLVM streams are still numbered 18, 19, and so on. It now follows
similar naming to GCC. It means 18.1 is not a version stream, but
stable tag prefix. THe 18.0.x releases are Alpha/Beta/RCs.

Note the package was still configured to use /llvm18/ paths
everywhere, and lots of related packages were missing.

Rename llvm18.1 back to llvm18, and add missing lld, cmake, clang
packages for version stream 18.

Also switch to using /llvm-18/ directories everywhere, like it is
common on all other operating systems.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@chainguard.dev>