calico: clean ups and fix ups
Make improvements and consolidate changes with versioned and fips
packaging of calico, to ensure these builds are easier to maintain in
the future.

When using go/build pipeline '-w -s' settings are already
pre-selected, and do not need to be specified again.

Also osusergo,netgo already results in static binaries with regular go
toolchain, thus forcing cgo off is not needed.

There are many other redundant things to clean up here, but this is
already a tangible improvement.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@chainguard.dev>