Update py/pip-build-install pipeline to work without specifying python
The pipelines/py/pip-build-install.yaml required 'python' to be set.
But if the caller was simply using 'python3', then there was no
way for the caller to know which 'pythonX.Y' that that would resolve
to.

Simply defaulting to 'python3' would result in using 'python3' as the
shbang for things that pip installed which is wrong, because if
python3 symlink changed between build-time and package install time
(which could be months later), runtime would be broken.