gcc: stop building go and objc
gccgo only supports EOL golang 1.18 without generics, it is thus not
useful to build up to date golang code. Unused by any of the packages
in Wolfi and not shipped in any Chainguard Images. It is holding up
shipping EOL go-1.18 in Wolfi.

objc compiler typically requires gnustep libraries to be usable, which
Wolfi does not package. Unused by any of the packages in Wolfi and not
shipped in any Chainguard Images.

These compilers can be reintroduced if popular software starts to
require these toolchains, or if there is demand to ship them as
Chainguard Images. In the meantime, building these toolchains is a
waste of resources.