gcc-12: add C and C++ tests
A recent update of gcc-12 resulted in incorrect runtime SCA
dependencies which caused compilers to not work at all. Add tests that
exercise both frontend and underlying compiler binaries to ensure they
are operation and are able to compile C and C++ code.

This is a smoketest, but would have caught the breakage in gcc-12
update.

Epoch bump not needed for test only change, as presubmit will rebuild
and tests will be executed against current builds.