chore(gitlab): Migrate packages with nonstandard releases to gitlab-cng
This migrates the GitLab container registry, ElasticSearch indexer, exporter,
logger, mailroom, and shell to GitLab CNG so that we can leverage the manifest
to build packages that align with the current GitLab release. This ensures that
GitLab components are not mismatched

This involves ugly workarounds. For example, we can not use our git checkout
pipeline and pass the component version so we just directly invoke git. A
few go/build and ruby/install pipelines also had to be expanded to workaround
the inability for passing the proper package version

Unfortunately, this is the best solution we have for tracking the in-use versions
of all of these different components

GitLab KAS, runner, and pages all follow the release structure of GitLab so those
have been left intact as individual packages

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>