gcc-12: config - fix false invalid march extension crc
Suspecting that the minimal patch is just:

    -|| [[ -n $opt_line ]]; then
    +|| [ x"$opt_line" != x ]; then

But cherry-picking the whole thing as it applies cleanly, and is full
provenance. Note all of config is compiler internal configuration as
used at compiler build time only and not present in the resulting
installed compiler.

This is to enable our default march setting as a built-in default.