py3-pip - drop flit-core as a runtime dependency.
Pip itself does not have any imports of flit or flit_core.
This was added in the pyversions refactor probably because
of a misunderstood dependency.  Several packages have
a pyproject entry like wheel does:

   pyproject.toml:requires = ["flit_core >=3.8,<4"]
   pyproject.toml:build-backend = "flit_core.buildapi"

And a simple 'pip build' will show a stack trace that doesn't
actually even leave pip so it appears/appeared that pip
had a flit_core dep.

Reality is that flit-core should be in the build environment
of those packages.