Drop flit-core from pip deps. (#24445)
* py3-pip - drop flit-core as a runtime dependency.

Pip itself does not have any imports of flit or flit_core.
This was added in the pyversions refactor probably because
of a misunderstood dependency.  Several packages have
a pyproject entry like wheel does:

   pyproject.toml:requires = ["flit_core >=3.8,<4"]
   pyproject.toml:build-backend = "flit_core.buildapi"

And a simple 'pip build' will show a stack trace that doesn't
actually even leave pip so it appears/appeared that pip
had a flit_core dep.

Reality is that flit-core should be in the build environment
of those packages.

* Add flit-core to dependencies of sepveral py3 packages.

These py3 packages were building because pip had flit-core
in its dependencies.  They need to explicitly add it.

With this change I can build py3-* that had either python-supported-pip
or py3-pip.

* Drop py3-pip-noflit

This package was only added to provide pip without flit as a dependency.
Flit-core is now removed from pip dep so it is not needed.