fix py3-gcovr (#24499)
* Add py/pip-check pipeline for running 'pip check'

This is most useful for packages that we have manually
encoded runtime package dependencies.  Running 'pip check'
will verify that the versions we have match expectation.

* py3-gcovr - fix missing dep, add tests

* Drop pip-check pipeline.

Discussed with pnasrat. There are benefits to running pip check,
but we'd like to do that more centrally (melange SCA?) or
correctly determine / generate runtime deps.