bash - fix test.
The test that was here had incorrect OR ('||')
when what was intended was PIPE ('|'):

    bash --version || grep "${{package.version}}"

Will run bash --version and only execute grep if it failed.

Changed to:

    out=$(/bin/bash --version)
    echo "$out" | grep "${{package.version}}"

This way , due to `set -e` checks that 'bash --version'
exits success and that its stdout has package.version inside.

Also improve the second test to validate the output rather than
only checking that bash exited zero.