Add py3-pip-wheel and py3-setuptools-wheel as for ensurepip
cpython's stdlib package 'ensurepip' bundles a pip wheel inside
the python distribution.  That is used to 'ensure' that there
is a pip inside a `python3 -m venv mydir`.  In python 3.8 to
python 3.11 there is also a bundled setuptools wheel.

That wheel is updated occasionally.

In order to make sure we're shipping the most recent versions
of setuptools and pip, we are removing the wheels from the
python packages and adding a dependency on these "wheel"
packages.