wasker: Fix wasker builds (#26189)
rust-1.73 was updated to require llvm17 which is bringing llvm into the
build env. llvm17+ packages conflict with the llvm-15 packages. wasker
the is specifying llvm15 as an explicit build dependency for inkwell. In
theory we could update the inkwell feature to build for llvm18 instead
but then we'll fail to build every time we update rust + llvm or pin it
to a specific version anyway and we have to update the projects cargo
deps so we'll just pin it to the llvm version upstream is using anyway.

<!---
Provide a short summary in the Title above. Examples of good PR titles:
* "ruby-3.1: new package"
* "haproxy: fix CVE-2014-123456"
-->

<!--
Please include references to any related issues or delete this section
otherwise.
 -->

Fixes:

https://github.com/wolfi-dev/os/issues/23864

Co-authored-by: Justin Vreeland <justin.vreeland@chainguard.com>