perl: build with openssf-compiler-options enabled
Perl builds currently is missing multiple hardening features, with
openssf-compiler-options many of them get enabled, specifically
bind-now, control flow protection. And for most, but not all, binaries
also stack protector and fortify source.

Whilst not ideal in terms of enforcement, this is improvement over
status quo.

Further investigation of perl configuration flags (or patches) is
required, as perl builds in other distributions appears to have stack
protected for more binaries. But it could be a false negative.

Other fixes:
- Remove duplicate configure flag.
- Ensure dependent source code is generated first, to prevent parallel make failures.
- Ensure code is writable, for strip to process binaries.
- Add hardening-check pipeline.
- Fixup hardening-check pipeline for tiny files, smaller than ELF magic.