protobuf-c: Fix riemann-c-client build (wolfi-dev/#23359) (#26208)
It looks like we pulled in an unmerged patches [1] from which is now
causing build issues with riemann-c-client. Once the patch causing
issues is removed riemann-c-client builds cleanly again.

https://github.com/protobuf-c/protobuf-c/pull/711

Fixes:
https://github.com/wolfi-dev/os/issues/23359


PR:

I just downloaded the patch from the GitHub PR request instead of
applying it and using format patch so some of the top matter is gone
now. I can restore it if that's an issue.

Notes:

We should validate this doesn't break existing builds before merging.

Co-authored-by: Justin Vreeland <justin.vreeland@chainguard.com>