gitaly-17.2: withdraw incorrect version 17.3 (#26474)
I have another PR
https://github.com/wolfi-dev/os/pull/26468#issuecomment-2293279759 which
reverts the version from the package yaml and fixes the update. This is
failing, I think because the bad 17.3 version is in the index and
getting installed in the test env which fails that presubmit check.


FWIW when I looked in the apk repo I could see the abseillib apks that
the withdrawn-packages.txt should have removed. That's why I've left
those in this change but removed the others, as they have been
withdrawn.

Signed-off-by: James Rawlings <jrawlings@chainguard.dev>