FTBFS collectd (#26484)
Fix failure to build from source of collectd. Update environment
dependencies on the dev package for varnish, and ask to link with
libgcrypt harder as configure.ac is not using pkg-config and thus does
not set GCRYPT_LIBS=-lgcrypt. Easier to do this as a variable, than
patching configure.ac.