bump patch release (#27044)
Bump py3-setuptools package to pick-up fix for:
https://github.com/scipy/scipy/issues/21416

This was a bug introduced upstream which is causing a number of our
packages to fail to build. This patch release resolves it:
https://github.com/pypa/setuptools/commit/ebddeb36f72c9d758b5cc0e9f81f8a66aa837d96

The tag was cut 4 days ago but they haven't cut a release yet - we're
monitoring releases not tags. But given this is causing package
failures, bumping ahead.

Signed-off-by: Mark McCormick <mark.mccormick@chainguard.dev>