prometheus: disable net/http/pprof (#26959)
Whilst https://pkg.go.dev/net/http/pprof is a useful diagnostics tool,
it exposes a potentially unaaccounted for endpoint. It is unsuitable to
have this enabled in production deployments.

Compile out this functionality. Ideally there would be a toolchain build
tag to turn this functionality off
https://github.com/golang/go/issues/69030 until then sed and patch it
out by hand.