unzip: fix build by building with linux_noasm
The generic targets adds -DNO_DIR and we do in fact have the headers for DIR.
Building just for linux pulls in some i386 asm code and the instructions
say for non intel platforms use no_asm.  In the future we should look
into using a custom set of flags and importing other distros fixes.