openssf-compiler-options: redesign gcc integration
When no specs are specified gcc self-creates "built-in" default specs
in memory.

If /usr/lib/gcc/arch/version/specs exists, it loads that, instead of
generating "built-in" default specs. And cruitially does not generate
"built-in" default specs at all.

Specs passed in via "-specs=<file>" command are looked up in the same
paths as "specs", however gcc still generates "built-in" default
specs.

Also /usr/lib/gcc/arch/version/specs file is not allowed to use
include commands, and is thus must be static / fully
self-contained. The general recomendation is to start with `gcc
-dumpspecs` tweak that, and keep that updated. But that is very error
prone, as gcc default specs have a lot of boilerplate that is compiler
configure flags dependant.

There is no filepath to drop in specs that are loaded the same way as
`-specs=<file>` are loaded (after default "built-in" ones are
generated, with include commands allowed).

The scary part is that incompete specs still cause compiler to sort of
work, but do result in 'gcc_s' not being linked in, hence the many
build-failures from the mass rebuilds with openssf-compiler-options
enabled. And that many software manages to build fine, especially
those with resiliant built systems that emitted fully qualified linker
commands to gcc. Lack of 'gcc_s' in the default link commands resulted
in cryptic build failures such as "unable to add library, undefined
symbol _Unwind_Rewind, DSO libgcc_s missing from command line" and the
like.

None of this documented in gcc documentation for [Spec Files], and is
only obvious from reading gcc source code. See
[this](https://github.com/gcc-mirror/gcc/blob/1a1f818c60636f219edd2cb93485aa08e1a36c88/gcc/gcc.cc#L8470)
line that loads specs, without built-in defaults and prevents using
"include" commands with the second argument set to true.

This is a very unfortunate set of circumstances. Patching compiler
specs loading is trivial in above linked code, but acceptance by
upstream is unknown at this point. I will attempt a submission and
will link it here at a later date. Thus resort to shell wrapper in a
higher priority path directory for now.

This should fix many build failures observed when trying to mass
enable openssf-compiler-options. One example is that cmake failed to
even configure its trivial bootstrap-cmake without this fix. Also
trivial package test added that shows problem with the current build,
which is resolved with wrapper solution.

[Spec Files]: https://gcc.gnu.org/onlinedocs/gcc/Spec-Files.html