binaryen - fix build with gcc-14
I was unable to get this to buidl without the provided patch.
Getting -Wno-error=maybe-uninitialized into the
compiler flags was not enough.

There are other examples in binaryen's git where this
pragma approach is used.

I could not get it to build with openssf-compiler-options
enabled, very close.  By adding

    vars:
      no-error-flags: -Wno-error=conversion \
        -Wno-error=float-conversion \
        -Wno-error=format-nonliteral \
        -Wno-error=maybe-uninitialized \
        -Wno-error=stringop-overflow

and then adding to the cmake/configure opts

  -DCMAKE_CXX_FLAGS="${{vars.no-error-flags}}"

we can get to an _Unwind_Resume linker error.
Presumably that will be fixed with something like
https://github.com/wolfi-dev/os/pull/27815