py3-pip - provide py3-pip-wheel
When we took the embedded wheel out of python3 packages (#25860) and
started depending on py3-pip-wheel, we could not use py3-pip package
because the py3-pip package requires python to build.  So we made
the py3-pip-wheel package which directly downloaded from pypa and
thus had no build dependency on python.

The fallout of that was that the py3-pip-wheel package was not being
patched (as seen in (#26220).

We could either
1. accept that we would have to patch the downloaded wheel
2. adjust things so we could use py3-pip origin.

Now py3-pip will build py3-pip-wheel.  No one should get the
py3-pip-wheel-bootsrap package version other than for bootstrapping.