krb5: improve configuration options (#28595)
Remove nonexistent without-tcl configure option that does nothing.

Previously TLS configuration was set to automatic, and could be
skipped if OpenSSL somehow becomes incompatible. Explicitely set to
OpenSSL such that it is guaranteed that OpenSSL support is not lost.

Previously built-in cryptography was used in libk5crypto.so.3, but
given openssl is in use for TLS, it is no additional dependency to
switch libk5crypto to use openssl as well.