kservice-modelmesh-serving: Fix internal config.
config/internal needs to be present with the binary, but the default
config gets mounted in the deployment yaml so doesn't need to be a part
of the package.

Q: Should the config be a separate subpackage/folder? Technically it can be in
a different location, but it always needs to be relative to the workdir
where the binary is being run from.