this commit adds 2207 simple tests across 625 packages (#28865)
This branch adds 2207 simple tests, across 625 packages that don't yet
have any tests. All of these tests were very simply generated by looking
for executables installed in the $PATH, which can take a common
parameter such as "--version" or "--help" or similar, and exits with
return code 0. Generally that should give us some regression protection
and signal that binaries installed continue to be executable... At
@joshrwolf request, I've added an # AUTOGENERATED comment inline, and I
did NOT bump the epochs (so as not to trigger a rebuild).

---------

Signed-off-by: Dustin Kirkland <kirkland@chainguard.dev>